renamed test_params to test_metadata
